refactor: made useTabs as hook
